Fix various crashes if debug libraries used (#3144)
* FuriHal: enable HSI in stop mode only if we use STOP0, proper SMPS selected clock assert

* Furi: fix double crash caused by bkpt use outside of debug session

* Libs: update ERC and MGG contrast

* Fix various crashes with LIB_DEBUG=1

* BadUsb: size_t where it should be and proper printf types

* Various fixes and make PVS happy

* FuriHal: proper CCID status and make PVS happy

* boot: update mode: graceful handling of corrupted stage file

---------

+++ applications/main/bad_usb/helpers/ducky_script.h
@@ -24,10 +24,10 @@ typedef enum {
 
 typedef struct {
     BadUsbWorkerState state;
-    uint16_t line_cur;
-    uint16_t line_nb;
+    size_t line_cur;
+    size_t line_nb;
     uint32_t delay_remain;
-    uint16_t error_line;
+    size_t error_line;
     char error[64];
 } BadUsbState;
